ir-alem/requirements.txt
paho-mqtt==1.6.1
flask==2.3.0
